requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
psycopg2-binary==2.9.5
python-decouple==3.6
sqlparse==0.4.3
